Fix image layout when using css-layout
Flipboard/react-canvas#16 describes an issue with transparent PNGs not
rendering correctly. This was solved in 0146a0b by introducing a separate
background layer, which would render a plain surface with the specified
background color and then putting the actual image (without a background color)
before this surface. However, when using css-layout/flexbox to layout
components, this background surface was considered another valid component
sharing the available space with the actual image. This leads to an undesired
result, described in Flipboard/react-canvas#43.

I fixed this issue by reducing the React render tree down to the image again,
but moved the background layer creation to the canvas actual drawing stage.

Fixes Flipboard/react-canvas#43.